Move utility functions to a separate class
